requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.4
django-cors-headers==3.13.0
django-model-utils==4.3.1
django-rest-framework==0.1.0
django-urlshortner==0.0.2
django-user-agents==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==22.0
Pillow==9.3.0
PyJWT==2.6.0
pytz==2022.6
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
segno==1.5.2
sqlparse==0.4.3
tzdata==2022.7
ua-parser==0.16.1
uritemplate==4.1.1
urllib3==1.26.13
user-agents==2.2.0
